Fix math bug
Hardcoded offset to account for a math bug. The old offset behavior was to take the first quaternion from the tracker, negate and save it, then multiply every following quaternion by the offset. This mostly worked but sometimes one of the values from the tracker would be exactly the same as it's corresponding value in the offset which would cause erroneous rotation values to be sent to slimevr, breaking stuff like drift correction. This commit hardcodes the offset to an imprecise value, ensuring that a quaternion from the tracker can't collide with the offset.